refactor: move the subject id param name config to the subject folder
